requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
attrs==23.2.0
certifi==2024.2.2
cloudflareai==0.4.0
discord.py==2.3.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
multidict==6.0.5
python-dotenv==1.0.1
sniffio==1.3.0
starlette==0.36.3
yarl==1.9.4
